+++ opensquirrel/passes/mapper/general_mapper.py
@@ -2,13 +2,15 @@
 
 from __future__ import annotations
 
+from typing import Any
+
 from opensquirrel.passes.mapper.mapping import Mapping
 
 
 class Mapper:
     """Base class for the Mapper pass."""
 
-    def __init__(self, qubit_register_size: int, mapping: Mapping | None = None) -> None:
+    def __init__(self, qubit_register_size: int, mapping: Mapping | None = None, **kwargs: Any) -> None:
         """Use ``IdentityMapper`` as the fallback case for ``Mapper``"""
         physical_qubit_register = list(range(qubit_register_size))
         self.mapping = mapping if mapping is not None else Mapping(physical_qubit_register)