+++ CHANGELOG.md
@@ -4,17 +4,20 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Types of changes:
+
 * **Added** for new features.
 * **Changed** for changes in existing functionality.
 * **Fixed** for any bug fixes.
 * **Removed** for now removed features.
 
 
-## [0.4.0] - [ xxxx-yy-zz ]
+## [ 0.4.0 ] - [ 2025-04-28 ]
 
 ### Added
 
+- Assembly declaration
 - `Rn` unitary instruction
+- `SWAP2CZDecomposer` decomposer pass
 - `CZDecomposer` decomposer pass
 - `ShortestPathRouter` router pass
 - `RandomMapper` mapper pass
@@ -26,8 +29,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Refactor: removed generators
 - Changed the `RoutingChecker` pass to a `RoutingValidator` pass
 - Changed use of `native` to `primitive`, e.g. `NativeGateValidator` is now `PrimitiveGateValidator`
+- Compilation passes accept `kwargs` as input
 
-## [0.3.1] - [ 2025-01-31 ]
+## [ 0.3.1 ] - [ 2025-01-31 ]
 
 ### Fixed